# api/openapi.yaml
openapi: 3.1.0
info:
  title: Pantry API
  version: 0.1.0
  description: |
    Legacy JSON/HTTP contract retained while Pantry adopts the Protobuf contract
    in proto/pantry/v1 over ConnectRPC. Health and readiness stay JSON-only.
    Shopping and scrape capabilities remain on their existing client paths until
    each has passed the compatibility gates in contracts/compatibility.md.
paths:
  /healthz:
    get:
      responses:
        '200':
          description: Process is live.
  /readyz:
    get:
      responses:
        '200':
          description: Foundation process is configured and accepting traffic.
  /api/v1/whoami:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verified caller identity.
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/v1/households:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Caller-visible households, authorized by existing Supabase RLS policies.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '502':
          description: Supabase REST was unavailable or rejected the request.
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseholdRequest'
      responses:
        '201':
          description: Household and caller owner membership were created atomically.
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '502':
          description: Supabase REST was unavailable or rejected the request.
  /api/v1/household-joins:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinHouseholdRequest'
      responses:
        '200':
          description: Caller joined using a valid invite, or was already a member.
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          description: Invite code was invalid, expired, or revoked.
        '502':
          description: Supabase REST was unavailable or rejected the request.
  /api/v1/membership:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The caller's first visible household membership, authorized by existing Supabase RLS policies.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '502':
          description: Supabase REST was unavailable or rejected the request.
  /api/v1/recipes:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRecipeRequest'
      responses:
        '201':
          description: Recipe and ingredient rows were saved atomically.
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '502':
          description: Supabase REST was unavailable or rejected the request.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthenticated:
      description: Missing, invalid, expired, or wrong-role access token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InvalidRequest:
      description: The request failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    CreateHouseholdRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        display_name:
          type: string
    JoinHouseholdRequest:
      type: object
      required: [invite_code]
      properties:
        invite_code:
          type: string
        display_name:
          type: string
    RecipeIngredient:
      type: object
      required: [name, quantity, unit, raw_string]
      properties:
        name:
          type: string
        quantity:
          type: [number, 'null']
        unit:
          type: string
        raw_string:
          type: string
    SaveRecipeRequest:
      type: object
      required: [household_id, title, ingredients]
      properties:
        household_id:
          type: string
        title:
          type: string
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RecipeIngredient'
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string